[#293] reverts commits for now to fix the build
The issue is we are using the raw response input stream which
hasn't been modified by the default middleware. This means if
the CI server returns a gzip response (as GitHub does when
returning the static XML we use in the functional test in
the build pipeline) it can't be parsed as it hasn't been
decoded.